Conflicts:
	app/Controller/PeopleController.php
	app/View/Layouts/default.ctp
	www/assets/fonts/ionicons.svg
	www/assets/fonts/ionicons.woff
	www/assets/fonts/ionicons.woff2

+++ app/Controller/HistoriesController.php
@@ -23,7 +23,7 @@ class HistoriesController extends AppController {
  */
 	public function index() {
 		$this->History->recursive = 0;
-		$this->set('histories', $this->Paginator->paginate());
+		$this->set('histories', $this->paginate());
 	}
 
 /**
@@ -35,7 +35,7 @@ class HistoriesController extends AppController {
  */
 	public function view($id = null) {
 		if (!$this->History->exists($id)) {
-			throw new NotFoundException(__('Invalid history'));
+			throw new NotFoundException(__('Historia No Existe'));
 		}
 		$options = array('conditions' => array('History.' . $this->History->primaryKey => $id));
 		$this->set('history', $this->History->find('first', $options));
@@ -50,15 +50,14 @@ class HistoriesController extends AppController {
 		if ($this->request->is('post')) {
 			$this->History->create();
 			if ($this->History->save($this->request->data)) {
-				$this->Session->setFlash(__('The history has been saved.'));
-				return $this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('La Historia Fue Registrada Correctamente'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
 			} else {
-				$this->Session->setFlash(__('The history could not be saved. Please, try again.'));
+				$this->Session->setFlash(__('La Historia No Fue Registrada Correctamente. Por Favor, Intente de Nuevo.'), 'flash/error');
 			}
 		}
 		$people = $this->History->Person->find('list');
-		$antecedents = $this->History->Antecedent->find('list');
-		$this->set(compact('people', 'antecedents'));
+		$this->set(compact('people'));
 	}
 
 /**
@@ -69,15 +68,16 @@ class HistoriesController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
+        $this->History->id = $id;
 		if (!$this->History->exists($id)) {
-			throw new NotFoundException(__('Invalid history'));
+			throw new NotFoundException(__('Historia No Existe'));
 		}
-		if ($this->request->is(array('post', 'put'))) {
+		if ($this->request->is('post') || $this->request->is('put')) {
 			if ($this->History->save($this->request->data)) {
-				$this->Session->setFlash(__('The history has been saved.'));
-				return $this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('La Historia  Fue Editada Correctamente'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
 			} else {
-				$this->Session->setFlash(__('The history could not be saved. Please, try again.'));
+				$this->Session->setFlash(__('La Historia  No Fue Editada Correctamente. Por Favor, Intente Nuevamente .'), 'flash/error');
 			}
 		} else {
 			$options = array('conditions' => array('History.' . $this->History->primaryKey => $id));
@@ -92,20 +92,23 @@ class HistoriesController extends AppController {
  * delete method
  *
  * @throws NotFoundException
+ * @throws MethodNotAllowedException
  * @param string $id
  * @return void
  */
 	public function delete($id = null) {
+		if (!$this->request->is('post')) {
+			throw new MethodNotAllowedException();
+		}
 		$this->History->id = $id;
 		if (!$this->History->exists()) {
 			throw new NotFoundException(__('Invalid history'));
 		}
-		$this->request->allowMethod('post', 'delete');
 		if ($this->History->delete()) {
-			$this->Session->setFlash(__('The history has been deleted.'));
-		} else {
-			$this->Session->setFlash(__('The history could not be deleted. Please, try again.'));
+			$this->Session->setFlash(__('History deleted'), 'flash/success');
+			$this->redirect(array('action' => 'index'));
 		}
-		return $this->redirect(array('action' => 'index'));
+		$this->Session->setFlash(__('History was not deleted'), 'flash/error');
+		$this->redirect(array('action' => 'index'));
 	}
 }
